Basic knowledge representation in state
World switches randomly between rewarding with A boxes, B boxes, or both. The agent does not know which world it is in until it visits an A box and receives no reward (or vice versa). The optimal policy, which the agent is able to learn, is to visit the closest box of either type, then visit the closest box of the other type if no reward is given.